implemented and tested a integrity check for the arguments of the convolutional layer

+++ ConvolutionalNeuralNetwork/test_result.hpp
@@ -4,7 +4,7 @@
 class test_result
 {
 public:
-	int data_count = 0;
+	size_t data_count = 0;
 	long long time_in_ms = 0;
 	float avg_cost = 0;
 	float accuracy = 0;
